+++ src/models/Settings.php
@@ -10,23 +10,10 @@
 
 namespace nystudio107\instantanalytics\models;
 
-use nystudio107\instantanalytics\InstantAnalytics;
-
-use Craft;
 use craft\base\Model;
 use craft\validators\ArrayValidator;
 
 /**
- * InstantAnalytics Settings Model
- *
- * This is a model used to define the plugin's settings.
- *
- * Models are containers for data. Just about every time information is passed
- * between services, controllers, and templates in Craft, it’s passed via a
- * model.
- *
- * https://craftcms.com/docs/plugins/models
- *
  * @author    nystudio107
  * @package   InstantAnalytics
  * @since     1.0.0
@@ -160,16 +147,8 @@ class Settings extends Model
     // Public Methods
     // =========================================================================
 
+    /** @noinspection ReturnTypeCanBeDeclaredInspection */
     /**
-     * Returns the validation rules for attributes.
-     *
-     * Validation rules are used by [[validate()]] to check if attribute values
-     * are valid. Child classes may override this method to declare different
-     * validation rules.
-     *
-     * More info:
-     * http://www.yiiframework.com/doc-2.0/guide-input-validation.html
-     *
      * @return array
      */
     public function rules()