+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Instant Analytics Changelog
 
-All notable changes to this project will be documented in this file.
+## 1.0.7 - 2018.08.24
+### Added
+* Added welcome screen after install
+* Automatically set the `documentTitle` from the SEOmatic `<title>` tag, if SEOmatic is installed
+* Automatically set the `affiliation` from the SEOmatic site name, if SEOmatic is installed
+
+### Changed
+* Lots of code cleanup
+* Moved to a modern webpack build config for the AdminCP
+* Added install confetti
 
 ## 1.0.6 - 2018.03.22
 ### Added